[bug] Fixed issue of interrogating private fields.
When the `renderer` encounters a `struct` it will attempt
to apply it's self on the fields of that struct. If the field
is private and implements the renderer interface, this was
causing a panic, as we were attempting to access a private
field from outside of the package.

We now skip over private fields.